refactor: data 디렉터리 remote, local 분리 작업
