+++ lib/data/remote/model/error_report_response.dart
@@ -1,5 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:santaclothes/data/model/care_label_details_response.dart';
+
+import 'care_label_details_response.dart';
 
 part 'error_report_response.g.dart';
