+++ lib/data/repository/login_repository.dart
@@ -1,6 +1,6 @@
-import 'package:santaclothes/data/model/token_response.dart';
-import 'package:santaclothes/data/network/auth_provider.dart';
-import 'package:santaclothes/data/prefs/token_manager.dart';
+import 'package:santaclothes/data/local/prefs/token_manager.dart';
+import 'package:santaclothes/data/remote/model/token_response.dart';
+import 'package:santaclothes/data/remote/network/auth_provider.dart';
 
 class LoginRepository {
   Future<void> postAuthRegister(
