+++ lib/data/repository/onboarding_repository.dart
@@ -1,4 +1,4 @@
-import 'package:santaclothes/data/prefs/onboarding_manager.dart';
+import 'package:santaclothes/data/local/prefs/onboarding_manager.dart';
 
 class OnboardingRepository {
   Future<bool?> getIsGuide() => OnboardingManager.instance.getIsGuide();
